refactor(api): [login] move fastify instance decorators
